requirements.txt
motor~=2.5.0
python-dotenv~=0.17.1
aiohttp~=3.7.4.post0
imageio~=2.9.0
coloredlogs~=15.0.1
pandas~=1.3.1
APScheduler~=3.7.0
discord.py==1.7.3
pymongo[srv]~=3.12.0
